requirements.txt
streamlit
faiss-cpu
tiktoken
openai
youtube-transcript-api
pandas
numpy
nltk